requirements.txt
Click==7.0
requests==2.22.0
PlexAPI==3.2.0
PyYAML==5.1.2
